Reorganized the structure of layout module
